pyproject.toml
[project]
name = "projeect-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pdf2image",
    "pytesseract",
    "pillow",
    "pymupdf>=1.27.2.2",
    "weasyprint>=68.0",
    "reportlab>=4.4.10",
    "opencv-python>=4.13.0.92",
]
